backend/requirements.txt
blinker
monai[all]
nibabel
nilearn
torchinfo
gdown
matplotlib
fastapi
uvicorn
pydantic
google-cloud-storage
requests
